requirements.txt
fastapi>=0.109.0
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==1.0.3
pydantic>=2.6.1
python-multipart==0.0.6
httpx>=0.23.0,<0.24.0
aiohttp>=3.8.0
requests>=2.31.0
huggingface_hub>=0.19.3
transformers==4.37.2
tokenizers==0.15.2

--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu 